requirements.txt
streamlit==1.35.0
openai==1.30.1
requests==2.31.0
python-dotenv==1.0.1   # (Optional: for managing API keys via .env) 